app/requirements.txt
Flask

mysql-connector